refactor: move wpml.py to output/dji.py
