+++ src/backend/packages/drone-flightplan/drone_flightplan/__init__.py
@@ -5,7 +5,7 @@ from drone_flightplan.calculate_parameters import calculate_parameters
 from drone_flightplan.create_placemarks import create_placemarks
 from drone_flightplan.waypoints import create_waypoint
 from drone_flightplan.add_elevation_from_dem import add_elevation_from_dem
-from drone_flightplan.wpml import create_wpml
+from drone_flightplan.output.dji import create_wpml
 from drone_flightplan.create_flightplan import create_flightplan
 
 __all__ = [
