+++ src/backend/packages/drone-flightplan/drone_flightplan/create_flightplan.py
@@ -9,8 +9,8 @@ from drone_flightplan.add_elevation_from_dem import add_elevation_from_dem
 from drone_flightplan.calculate_parameters import calculate_parameters
 from drone_flightplan.create_placemarks import create_placemarks
 from drone_flightplan.waypoints import create_waypoint
-from drone_flightplan.wpml import create_wpml
 from drone_flightplan.drone_type import DroneType
+from drone_flightplan.output.dji import create_wpml
 
 # Instantiate logger
 log = logging.getLogger(__name__)
